# [aiohttp] - disable reset query and remove explicit call prepare