@Reskov 2025-04-16
@Dreamsorcerer please take a look. Performance almost twice better for `/queries` endpoint on my local machine.