@Dreamsorcerer 2025-04-16
> @Dreamsorcerer please take a look. Performance almost twice better for `/queries` endpoint on my local machine.

If there's such a big difference, I wonder if something should be changed upstream? Looking through the code, it looks like it already skips that if the server doesn't have those capabilities:
I wonder if the same result can't be achieved by disabling some server capabilities?
https://github.com/MagicStack/asyncpg/blob/5b14653e0b447d956aa01ec658562138e19f0293/asyncpg/connection.py#L1747-L1754

If it's CrateDB or CockroachDB then it skips all of that anyway, and on Amazon RedShift only does the RESET ALL. Only on standard PostgresQL does it run all of that query:
https://github.com/MagicStack/asyncpg/blob/5b14653e0b447d956aa01ec658562138e19f0293/asyncpg/connection.py#L2661-L2697

Which makes me wonder if it's really necessary after all...