@Dreamsorcerer 2025-04-16
> I do not think that disable only `unlisten` will make a noticeable difference. The big difference is to execute or not whole reset query. Some of the options like parameter set for `RESET ALL`, looks not so easy to be detected.

If you try hardcoding the query locally, can you verify if that's true? My suspicion is that with the async nature, that maybe the contents of the query is having an impact (there are 4 queries in there after all). I'm a little surprised to see double the performance for that, as it suggests that the reset query is actually taking more time than the regular queries (given that the DB query can't be 100% of the benchmark time).