@Dreamsorcerer 2025-04-16
> If you try hardcoding the query locally

So, hardcode to:
```
return 'SELECT pg_advisory_unlock_all();CLOSE ALL;UNLISTEN *;RESET ALL;'
```
Then remove each statement and verify how much impact each part of the query has.