@Reskov 2025-04-16
This is hard to check locally since a have a large deviation due to I run tests on virtual machine and low test duration. So if we want to know exact percentage of improvement. We have to prepare bare-metal machine and test only asyncpg without nginx/aiohttp overhead.


### SELECT pg_advisory_unlock_all();CLOSE ALL;UNLISTEN *;RESET ALL;
```
 Queries: 1 for query
 wrk -H 'Host: tfb-server' -H 'Accept: application/json,text/html;q=0.9,application/xhtml+xml;q=0.9,application/xml;q=0.8,*/*;q=0.7' -H 'Connection: keep-alive' --latency -d 30 -c 32 --timeout 8 -t 6 "http://tfb-server:8080/queries/1"
---------------------------------------------------------
Running 30s test @ http://tfb-server:8080/queries/1
  6 threads and 32 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.05ms  708.96us  35.15ms   86.89%
    Req/Sec     5.01k     1.03k    6.28k    82.28%
  Latency Distribution
     50%    0.89ms
     75%    1.29ms
     90%    1.79ms
     99%    3.37ms
  896495 requests in 30.00s, 159.69MB read
Requests/sec:  29880.08
Transfer/sec:      5.32MB
STARTTIME 1744814199
ENDTIME 1744814229
```
### Select 1

```
class NoResetConnection(asyncpg.Connection):
    __slots__ = ()

    def get_reset_query(self):
        return 'SELECT 1'
```


```
  Queries: 1 for query
 wrk -H 'Host: tfb-server' -H 'Accept: application/json,text/html;q=0.9,application/xhtml+xml;q=0.9,application/xml;q=0.8,*/*;q=0.7' -H 'Connection: keep-alive' --latency -d 30 -c 32 --timeout 8 -t 6 "http://tfb-server:8080/queries/1"
---------------------------------------------------------
Running 30s test @ http://tfb-server:8080/queries/1
  6 threads and 32 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.91ms  530.35us  15.76ms   81.01%
    Req/Sec     5.67k   418.48    13.11k    81.34%
  Latency Distribution
     50%  788.00us
     75%    1.13ms
     90%    1.54ms
     99%    2.62ms
  1015678 requests in 30.10s, 180.91MB read
Requests/sec:  33744.19
Transfer/sec:      6.01MB
STARTTIME 1744814374
ENDTIME 1744814405
```


```
class NoResetConnection(asyncpg.Connection):
    __slots__ = ()

    def get_reset_query(self):
        return ''
```
```
---------------------------------------------------------
 Queries: 1 for query
 wrk -H 'Host: tfb-server' -H 'Accept: application/json,text/html;q=0.9,application/xhtml+xml;q=0.9,application/xml;q=0.8,*/*;q=0.7' -H 'Connection: keep-alive' --latency -d 30 -c 32 --timeout 8 -t 6 "http://tfb-server:8080/queries/1"
---------------------------------------------------------
Running 30s test @ http://tfb-server:8080/queries/1
  6 threads and 32 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   674.03us  382.59us  10.58ms   78.08%
    Req/Sec     7.64k   403.10    12.09k    76.37%
  Latency Distribution
     50%  588.00us
     75%  842.00us
     90%    1.15ms
     99%    1.87ms
  1370230 requests in 30.10s, 244.07MB read
Requests/sec:  45524.99
Transfer/sec:      8.11MB
STARTTIME 1744814546
ENDTIME 1744814576

```