@Reskov 2025-04-16
> I'm a little surprised to see double the performance for that, as it suggests that the reset query is actually taking more time than the regular queries (given that the DB query can't be 100% of the benchmark time).

Yeah, it is actually not double. Initially I have two optimisations inside this PR (remove call of prepare and remove reset query). Let's assume (1-33744.19/40608.53) around 17%